build(engine): migrate runtime to SDL3 and vendor SDL
Impact:
- Space now runs on SDL3 with no SDL2 compatibility path.
- CEF startup no longer crashes with ICU FD errors when browser surfaces initialize.
- Default test runs now exercise CEF initialization instead of bypassing it.

What:
- Vendored upstream SDL3 source into `external/sdl/` and wired build usage to SDL3.
- Migrated engine/window/input/timer/event handling from SDL2 APIs to SDL3 APIs.
- Renamed game controller code paths to gamepad naming (`input_game_controller_state` -> `input_gamepad_state`).
- Updated Lua/Fennel bindings/tests and refreshed affected E2E snapshot goldens.
- Fixed Linux CEF runtime payload staging (`icudtl.dat` and related resources) and hardened subprocess switches.
- Removed `SPACE_SKIP_CEF=1` from CTest properties for `space_fnl_tests` and `test_dotenv_integration`.
- Added a dev note documenting the ICU crash root cause and fix.

Why:
- Repository packages did not provide the required SDL3 setup for this project.
- A strict cutover was requested, prioritizing long-term cleanliness over dual support.
- CEF needed to be truly fixed and covered by normal test execution.

How:
- Integrated SDL3 as a vendored dependency in-tree, then migrated all call sites and tests to SDL3 semantics.
- Reproduced the CEF ICU crash, traced runtime lookup behavior, then staged resources where Chromium expects them on Linux.
- Re-ran full test suites and snapshot flows until normal + E2E passed with CEF enabled by default.

Risk:
- Vendoring full upstream SDL increases repository size and upgrade surface.
- SDL3 behavioral differences may still cause subtle runtime/input/render variance across platforms.
- Snapshot updates reflect intentional rendering changes from migration, but may hide unintended small visual drift.

Testing:
- `make cmake` (pass)
- `make build` (pass)
- `SKIP_KEYRING_TESTS=1 XDG_DATA_HOME=/tmp/space/tests/xdg-data SPACE_DISABLE_AUDIO=1 SPACE_ASSETS_PATH=$(pwd)/assets make test` (pass)
- `SPACE_DISABLE_AUDIO=1 XDG_DATA_HOME=/tmp/space/tests/xdg-data SPACE_ASSETS_PATH=$(pwd)/assets make test-e2e` (pass)

Follow-ups:
- Consider pruning non-essential upstream SDL files later if repository size becomes a maintenance issue.
- Pin/document SDL upstream revision policy for future updates.

References:
- Refs: SDL3 migration guide
- Refs: docs/dev/notes/cef-icu-fd-crash.md

+++ src/timer.cpp
@@ -1,8 +1,8 @@
 #include "timer.h"
 
-unsigned int Timer::computeDeltaTime() {
+Uint64 Timer::computeDeltaTime() {
     frameStart = SDL_GetTicks();
-    unsigned int dt = frameStart - lastFrame;
+    Uint64 dt = frameStart - lastFrame;
     lastFrame = frameStart;
     return dt;
 }
@@ -13,4 +13,3 @@ void Timer::delayTime() {
         SDL_Delay(frameDelay - frameTime);
     }
 }
-
